chore: move data url parse into rolldown_utils
