+++ crates/rolldown_plugin_data_url/Cargo.toml
@@ -9,7 +9,6 @@ doctest = false
 
 [dependencies]
 base64-simd     = { workspace = true }
-nom             = { workspace = true }
 rolldown_common = { workspace = true }
 rolldown_plugin = { workspace = true }
 rolldown_utils  = { workspace = true }
