+++ crates/rolldown_plugin_data_url/src/data_url_plugin.rs
@@ -5,9 +5,10 @@ use rolldown_plugin::{
   HookLoadArgs, HookLoadOutput, HookLoadReturn, HookResolveIdArgs, HookResolveIdOutput,
   HookResolveIdReturn, Plugin, PluginContext,
 };
-use rolldown_utils::dashmap::FxDashMap;
-
-use crate::utils::{is_data_url, parse_data_url};
+use rolldown_utils::{
+  dashmap::FxDashMap,
+  dataurl::{is_data_url, parse_data_url},
+};
 
 #[derive(Debug)]
 pub struct ResolvedDataUrl {
