+++ crates/rolldown_plugin_data_url/src/lib.rs
@@ -1,5 +1,4 @@
 #[allow(clippy::manual_async_fn)]
 mod data_url_plugin;
-mod utils;
 
 pub use data_url_plugin::DataUrlPlugin;
