+++ crates/rolldown_utils/Cargo.toml
@@ -32,6 +32,7 @@ indexmap           = { workspace = true, features = ["serde"] }
 infer              = { workspace = true }
 memchr             = { workspace = true }
 mime               = { workspace = true }
+nom                = { workspace = true }
 oxc                = { workspace = true }
 oxc_index          = { workspace = true }
 phf                = { workspace = true }
